requirements.txt
pandas
praw
apache-airflow
numpy
pyyaml